Improve tests for pthreads+dylink
test_pthread_dlsym.c: Explicit syncing of code is not necessary in this
test. The libraries should be loaded in the same order on every thread
regardless.

test_pthread_dlsym.c: Be a little more hostile by avoiding pthread
constructs and just spinning.  This mean the explicit code sync is more
likely to actaully be needed, even in the face of syncronization happing
if futex_wait (which I'm porposing to add as followup).