Add type checking for `-s` settings
For almost all types there is single type that can be inferred from the
default value.

There are a few exceptions, where a single settings couple have more
than one type (for example PTHREAD_POOL_SIZE can be an integer or
string).  For now, I simply exclude those few settings from the type
checking.

For now, I've not tried to add any extra annotations to the settings
file.

We could possibly followup by adding explicit annotations and/or support
for polymorphic settings, but this change is big with with minimal
impact.

Fixes: #16440