Use appropriate types for syscall arguments. NFC
The net effect of this change simply that there are fewer i64 argument
in the syscall fuctions for wasm64.  For wasm32 this change has no
effect at all.

Unlike the linux syscall layer we are not bount to fixed interface
consisting only of word-sized (long) arguments.

For wasm64 we want to avoid the use of `long` in order to limit the
number of BitInt arguments on the JS side.  In this change I've modified
the arguments from all being `long` into three different classes:

- `int`: For 32-bit values - these are the same under
         wasm32 and wasm64.
- `intptr_t/size_t`: For pointer-sized things that are
                     64-bit only under wasm64.
- `uint64_t`: For `off_t` and similar that are always
              64-bit.

There is very little usage of always-64-bit/off_t in the syscall layer
and I already landed that part of the change in: #16590