Fix return code from __syscall_getcwd
The return code from this syscall is the number of bytes written, not
the buffer pointer.

As it happens the libc function works just as well either way, but i've
been looking to change the return of all our syscalls from long to int
and I noticed this was the only system that had a pointer-sized return
value (thus requiring i64 on wasm64).  Turns out it doesn't require this
after all.

To show that this is that we expect the linux syscall to return here is
how glibc uses the return value:

https://github.com/lattera/glibc/blob/895ef79e04a953cac1493863bcae29ad85657ee1/sysdeps/unix/sysv/linux/getcwd.c#L78-L84