[wasm64] convert BigInt<->Number based on library function `__sig` attributes
This change introduces and type to the `__sig` scheme that we using for
library functions: 'p'.  This letter signifies that the parameter or
return value is the width of a pointer which means it will differ
between wasm32 and wasm64.

We use this whenever a JS function receives or returns pointer type or a
size_t type.

For wasm32, we don't need to do anything since the incoming value is and
i32.  For wasm64 we convert the incoming BigInt value to an int53/double
at the JS/Wasm boundary.  At the moment no checking is done on the
values, which means values that don't precisely convert to JS double can
be corrupted/truncated.  As a followup, we could consider using the
checking conversion functions, or asserting in debug builds if the values
don't fit within this range.