OpenAL: fixed invalid behavior and crashes in multi-source functions (#16633)
OpenAL multi-source functions (alSourcePlayv, alSourceStopv, alSourceRewindv,
alSourcePausev) were not behaving as intended, e.g. alSourcePausev never paused
any sound, and alSourceStopv even crashed with Uncaught TypeError: src.bufQueue
is undefined.

It turned out that inside those functions setting the state of the provided list of sources
was actually applied on the source-ids instead of the source objects.

This was not the case for the single-source functions.

This change should also fix #13797.