Fix dyncall and dynamic linking
We don't compile with WASM_BIGINT so if there is a 64 bit integer
in the dynamic call, Emscripten tries to call a legalizer wrapper.
These legalizer wrappers are only generated in the main module so
sometimes a side module tries to call a wrapper that doesn't exist.
This adds an extra handler in that case.