[WasmFS] Stub ioctl()/terminal support (#16677)
We don't have TTY support in emscripten anyhow, so most of this just
does nothing. We also do not yet have a concept of TTYs in wasmfs,
so just recognize stdin/out/err as such. This is enough for basic things,
gets some tests passing, and is almost at parity with the old FS (it had
the ability to create new terminals and mark them as such).