[WasmFS] fcntl() (#16710)
Fixes #15970

Changes flags to be behind a lock on file table entries, as now they can
be modified.

This basically matches the JS FS state, except that it does not claim to
support byte-range locks. As wasmfs is meant to support multithreading
well it seems best to not lie about that and return an error for now, until
we implement that (which custom backends might eventually).