Fix prototype for _dlinit()
As of LLVM revision
https://github.com/llvm/llvm-project/commit/33d3fc4466479285121cbb1a62db249454da0bda
Clang will warn when using a function with no prototye and passing it arguments.
Note that `void _dlinit()` is not a prototype, but `void _dlinit(void)` is.
However in this case the function actually does have an argument.

Also fix the discrepancy in the arguments  on the JS side.