Have MemoryDirectory call out to the backend to create children
Since MemoryDirectory is re-used for multiple kinds of in-memory backend, it
cannot hard code `MemoryFile` into `createDataFile`. Instead, call out to the
backend to initialize a new file. Although the backend methods are meant for
creating fresh mount roots, for in-memory backends there is usually no different
between a fresh mount root and any other kind of fresh file, so it is reasonable
to use those methods here.

However, using MemoryDirectory and friends makes no sense for persistent
backends, so also update `Backend` to no longer default to using them.