[WasmFS] Refactor the backend Directory API (#16739)
Previously new entries were inserted into the file system by calling
`Backend::createDirectory` or `Backend::createFile`, etc. then passing the
result to `Directory::insertChild`. This was suboptimal because it forced `File`
objects to be created with no context and separately from the creation of the
underlying files. Backends would have to create some reasonable default `File`
and then later overwrite its state in the implementation of `insertChild`.

To avoid this spooky action at a distance, simplify the control flow, and reduce
statefulness in the API, replace `insertChild` with separate `insertDataFile`,
`insertDirectory`, and `insertSymlink` methods that both create the underlying
files and allocate and return the corresponding `File` objects.

But removing `insertChild` from the backend API makes it no longer possible for
a backend to contain a reference to a file on a different backend, a capability
we need at least for the root in-memory backend. To fix this (and also to
improve performance and also to fix an entire class of bugs stemming from having
multiple `File` objects corresponding to the same underlying file), add a
directory entry cache inspired by the dcache in Linux's [virtual file
system](https://github.com/torvalds/linux/blob/master/Documentation/filesystems/vfs.rst).

This directory entry cache lives in the base `Directory` class and is
manipulated solely by the `Directory::Handle` layer sitting between the syscall
implementations and the backends. A new method, `Directory::Handle::mountChild`,
can insert `File` objects into this cache without inserting them into the
underlying backend, allowing directories to contain children in different
backends without backends having to worry about how to represent such children.

Finally, add another method, `Directory::insertMove`, that implements an atomic
rename to replace the combination of `removeChild` followed by `insertChild`
that was previously used to implement rename. Although the previous call
sequence worked fine for the in-memory backend, it would have been extremely
awkward to support in any persistent backend and was in general not capable of
gracefully handling rename errors without data loss.

After all these updates to the directory API, there was one additional fix
necessary to keep the tests passing. Now that there is a unique `File`
corresponding to each underlying file in the Node backend, multiple `open`s of a
file will share an underlying file descriptor. But if a file is originally
opened for reading only, for example, and later opened for writing, writes will
fail because the underlying Node file descriptor was only opened for reading.

One possible solution would have been to always have the Node backend open files
for both reading and writing, but that would introduce spurious failures when
the user only has permissions for one or the other on the underlying file.
Instead, detect when the permission on the current underlying file descriptor
are insufficient and dynamically try to increase permissions on demand.

Resolves #16630.