Remove `ignore` flag from `makeGetValue` in strptime
The ignore flag mean that we don't do SAFE_HEAP checks on these
accesses.  However we don't use the ignore flag on other writes to `tm`
structures in other libc functions, and also this change doesn't causes
any of the wasm2s tests to fail, so I assert that this is no longer
needed and might be a fastcomp artifact.

This change is part of a larger effort to simplify
makeGetValue/makeSetValue by removing these extra/unused params/modes.