Bound the number of postMessages generated by proxying (#16824)
The proxying queue implementation previously generated a new postMessage
notification every time a task_queue became nonempty. This notification scheme
could produce an unbounded number of postMessage notifications when a worker
thread executed a queue in a loop without ever returning to the event loop to
process its notifications.

To avoid this effective memory leak, produce no more than one pending
postMessage notification at a time for each task_queue by tracking notifications
on the granularity of each task_queue, replacing the single js_refcount field on
the parent em_proxying_queue. As before with the refcount, it is possible that
the notification flag will need to be cleared after a thread has died when there
is no stack available. To accommodate this, use the task_queue pointer instead
of the em_proxying_queue pointer as the postMessage payload and clear it
directly from JS when processing notifications.

The protocol for clearing the flag satisfies two constraints. First, it must
ensure that no work is dropped when proxying new work races with processing the
notification for old work. Clearing the flag after processing the queue would
not work because new work could arrive after the processing but before the
clearing and neither get processed due to the previous notification nor trigger
a new notification. On the other hand, the other constraint is that the flag
must not be cleared until the queue can be freed, so it cannot be cleared before
processing the queue. To satisfy these contradictory constraints, use a two-step
protocol that allows new notifications to be generated as soon as the current
notification processing starts but also keeps the queue alive until processing
finishes.

Fixes #16767.