Add note on emscripten_set_main_loop with Wasm EH
When `simulate_infinite_loop` is true, `emscripten_set_main_loop` does
not work with the new Wasm EH because throwing a JS exception will cause
destructors in the Wasm stack frames to run. This can be fixed in future
by either by fixing the LLVM compilation by adding an additional `catch`
to every `catch_all` so that we can rethrow JS exceptions of this kind,
or possibly using the new stack switching proposal.

This PR adds a note that this features currently doesn't work with Wasm
EH.