[CI] Switch to the previous Chrome download URL
In a824211356 we added a "_new" suffix to the URL we use to download Chrome on
CI to switch to a newer version hosted in our cloud storage bucket. Now that
that new version has proven to work, switch back to the normal URL that now
points to the new version as well.