Temporarily set -Wno-pointer-sign for zlib build (#16906)
zlib's configure script assumes that test compiles failed if there is any output on stderr (rather than being based on exit status). Sometimes emscripten prints "building libc" or similar if it needs to build a library. This causes flaky failures in running the tests. See #16908 for details.
This PR works around the problem by disabling the warning, but doesn't fix the underlying problem.