[WasmFS][NFC] Simplify the async JS backend code (#16894)
Remove the CallbackState argument in favor of directly passing outparam pointers
for results and remove the callback function in favor of passing the
em_proxying_ctx and directly calling emscripten_proxy_finish on it. Also remove
the unnecessary `runtimeKeepalivePush` and `runtimeKeepalivePop` calls since the
`ProxyWorker` utility already takes care of keeping the runtime alive.