Update implementation of deduplication function
Old implementation is very slow. It has O(n^2) complexity.

With large input data, you can wait for the result for hours,
while the new implementation works in just a couple of seconds.

New implementation uses dict that is significantly faster - O(n).

Since we require python 3.6, that ordering of dictionaries is guaranteed
to be insertion order so we can use 'dict' here but not 'set'.