[WasmFS] Add wasmfs_before_preload hook to run code before preloading (#16967)
Preloaded files are added during startup, but sometimes the user wants to
load them under a custom backend and not the default. In the old FS, the
user could mount directories in a Module.preRun, but with WasmFS that
has to be done in compiled code. To allow that, this defines a hook that
users can implement:

wasmfs_before_preload

If that the symbol is defined then it is called right before preloading.

To make this work the file preloading code will now skip any directories
that already exist, instead of erroring on them.