Remove now-unneeded check in library_pthread.js. NFC
The race condition decribed in this comment is no longer possible
because the very fist statement `S2: thread ends, 'exit' message is sent
to MB` is no longer true.

When a non-detached thread exits it no longer sends any kind of message
back to the main thread. It simply sets it own `detach_state` to
`DT_EXITED`.