[WasmFS] Fix OPFS `getSize` when file is not opened
The `getSize` implementation in the OPFS backend previously depended on having
an AccessHandle to the corresponding file, but that's only the case for open
files. `getSize` can also be called on files that have not been opened, for
example in the implementation of the `stat` syscall. Fix the bug in this case by
using an alternative code path that gets the size of the file using the Blob
API, which does not require having an AccessHandle.