[EH] Rename format_exception.cpp
I think we need a separate file in libc++abi for 'our' EH, meaning both
Emscripten EH and Wasm EH. (I think they can be in the same file, as
long as we have `ifdef`s) This also adds a file header.

I think this method can be in that file, rather than having a separate
file for this. I named this `cxa_emscripten.cpp` because it's the name
 #16627 used and I hope we can put all Emscripten specific stuff there,
but maybe @sbc100 had another thought; if so let me know.