Don't call `__stdio_exit` when we want to keep the runtime alive
`__stdio_exit` should only be used in cases where the runtime is no
longer going to be used but `checkUnflushedContent` is only used in the
cases where EXIT_RUNTIME=0 (i.e. the runtime stays alive forever).

This was a bug introduced in #16130, which manifested in deadlocks when
other threads attempted to use the stdio streams since __stdio_exit
called `FFINALLOCK` which leaving them in a forever-locked state.