Remove `source-map` dependency from terser bundle (#17067)
When updating terser to 4.1.0 in #17035 I inadvertently introduced a
dependency on the node source-map module.  Rebuilding after patching
our terser fork removed this:

https://github.com/emscripten-core/terser/commit/b3ace3b63703d23316b224ef00dd32669369a948

This happens to be available during testing because it one of our dev
decpendencies but we not include it in our end-user/production
dependencies.

This kind of error is hard to find since the `node_modules` for our test
setup is the same as the node_modules for emscripten itself.  It migth
be worth attempting to tease them appart if this kind of thing ever
shows up again.