Initial Wasm Stack Switching support (#16779)
This uses the Promise API for Wasm Stack Switching as an alternative to
Asyncify. That is, instead of modifying the wasm like Asyncify does, we have
the VM pause and resume us. When we want to pause an import returns a
Promise, and when we want to resume we resolve the Promise, basically.
See:

https://github.com/WebAssembly/js-promise-integration/

There is more work to do, like getting all the tests to pass, but for now this
gets one Asyncify method working (emscripten_sleep) and one test for it,
and adds a code size test that shows the benefit here.

For now the API to use stack switching is ASYNCIFY=2, but we may want
a more user-friendly flag at some point.