Open final_js file using 'utf-8' encoding
There is final_js file. It can contain unicode characters.
But then it is read without specifying the encoding.

Such a case was found, which leads to UnicodeEncodeError: use
dynamic linking with '--pre-js' option.

Solution is to specify the correct encoding. A unit test was
also written for this case.