[EH] Make 'getExceptionMessage' work for Wasm EH (#17064)
This makes `getExceptionMessage` work for Wasm EH as well.

I am not very happy with the design myself, and I think this needs some
explaining in various parts. This became kind of long but please bear
with me. Suggestions are welcome.

---
First, this exports `__cpp_exception` tag. In async compilation, we
delay assignment of exported wasm functions to `Module['asm']` by using
`createExportWrapper`:
https://github.com/emscripten-core/emscripten/blob/1f6b136d75c9a6a89d07ed069b9cede80b39cebd/src/preamble.js#L677-L694
But tags are not functions and we can't use this tactic. So currently we
can't do this to use the tag simply by a variable `___cpp_exception`:
```js
var ___cpp_exception = Module["___cpp_exception"] = createExportWrapper("__cpp_exception"); // Doesn't work!
```

So I resorted to create `$getCppExceptionTag` method in JS library. The
JS exception library also provides `$getCppExceptionThrownObject`, which
essentially calls `getArg` for you.

---
The complexity of this PR partly stems from one discrepancy between Wasm
EH and Emscripten EH. In Emscripten EH, the actual thrown value by
`throw` is an address of the exception object created by
`__cxa_allocate_exception`:
https://github.com/emscripten-core/emscripten/blob/1f6b136d75c9a6a89d07ed069b9cede80b39cebd/src/library_exceptions.js#L215

But in Wasm EH, what we throw is an `WebAssembly.Exception` object. You
can use `getArg` method to get the actual value thrown. But even this
value is not the thrown object like Emscripten EH. In libc++abi, which
Wasm EH uses, what we actually throw is a pointer to something called
unwind header, represented by `_Unwind_Exception`:
https://github.com/emscripten-core/emscripten/blob/1f6b136d75c9a6a89d07ed069b9cede80b39cebd/system/lib/libcxxabi/src/cxa_exception.cpp#L284

In libc++abi, the memory layout is as follows:
https://github.com/emscripten-core/emscripten/blob/1f6b136d75c9a6a89d07ed069b9cede80b39cebd/system/lib/libcxxabi/src/cxa_exception.cpp#L26-L35

Changing the thrown value in libc++abi to the actual thrown object
pointer will be nontrivial, and this makes our library deviates a lot
from libc++abi, which can interfere with future upstreaming efforts.

And changing Emscripten's thrown value to that of Wasm EH's is hard too,
because Emscripten EH doesn't even have a concept of unwind header.

So what I ended up doing is make different versions of
`$getExceptionMessage` in the JS library. In Emscripten EH's version it
simply calls `__get_exception_message` we added in
`cxa_exception_emscripten.cpp`. In case of Wasm EH, it extracts the
value using `getArg` and calls `__get_exception_message` with it.
`__get_exception_message` will convert the given value from a thrown
object to an unwind header in case of Wasm EH.

---
The other discrepancy is the handling of refcount in Emscripten EH and
Wasm EH. As I wrote in the `FIXME` comment in `test_exception_message`,
Wasm EH sets the refcount to 1 when throwing, does nothing in
`__cxa_begin_catch`, and decrements it in `__cxa_end_catch`. But
Emscripten EH sets the refcount to 0 when throwing, increments it in
`__cxa_begin_catch`, and decrements it in `__cxa_end_catch`. I tried to
change Emscripten EH JS library to match Wasm EH's way, but it was
intertwined with other stuff and nontrivial. I think we eventually need
to do it, but I didn't want to delay this PR further because of that.

So the current hacky solution in this PR is to make common wrapper
functions `$incrementExceptionRefcount` and
`$decrementExceptionRefcount` in JS library. In Emscripten EH, it just
calls `__cxa_(in|de)crement_exception_refcount` in JS. In Wasm EH, we can't
call libc++abi's `__cxa_(in|de)crement_exception_refcount` directly
because as I described earlier, the thrown object is different; Wasm EH
throws an unwind header pointer. So I created another layer of wrappers,
`__(in|de)crement_wasm_exception_refcount` in
`cxa_exception_emscripten.cpp`, which converts an unwind header pointer
to a thrown object, and calls libc++abi's
`__cxa_(in|de)crement_exception_refcount`. This is the part I'm very not
happy about.

---
Currently this is not particularly easy to use. As you can see in
`test_exception_message`, you need to export several functions manually
to use this. Furthermore, after this patch, the set of functions you
need to export is different in Emscripten EH and Wasm EH. I think
crerating a setting for this with which all the necessary functions are
automatically exported would be a follow-up. But other suggestions are
welcome too.

---
This partly addresses #16033 and #16380. I'm not sure if this "fixes"
them because this is not very easy to use at this point, as I described
in the previous paragraph.

Companion patch: https://github.com/WebAssembly/binaryen/pull/4691