Fix signature generation in RELOCATABLE (#17139)
This fixes wasmfs.test_relocatable_void_function but the problem is not
specific to WasmFS. The bug was that we checked MAIN_MODULE where
we should check RELOCATABLE (really, we don't have much use case for
RELOCATABLE-with-JS-but-not-MAIN_MODULE, but that's a separate
issue). This was not noticeable before because the few tests we have for
RELOCATABLE by itself did not happen to have imports that actually need
a signature. But in WasmFS we have more imports, specifically we add the
__emscripten_out import to print to stdout, and that ends up crashing on
not having a signature.