Fix for weak undefined symbols in dynamic linking
With this change we maintain a global set of weak undefined symbols
(WeakSymbols), when modules are loaded the `dylink` section tells us
which imports are weak imports.

We then resolve weakly imported symbol addresses to zero without
generating any errors.

Sadly there is some asymmetry with the way the main module is loaded
and the way side modules are, which makes this patch a litte more
complex than I would like.  For example the weak imports of the main
module are extracted at build time rather than runtime.

Fixes: #12819