[NFC] Split optz functions out of libc (#17179)
Put them in a new libc_optz library that will override parts of libc with the
size-optimized versions instead. This avoids rebuilding all of libc for optz,
and can save a few minutes on build-linux on CI.

Avoid optz in MAIN_MODULE=1, because this override method (link in
libc_optz first so it overrides libc) conflicts with it. But that is fine, as in
MAIN_MODULE=1 the code size benefit of optz is not noticeable anyhow.

Add a test for MAIN_MODULE=1 + -Oz, which was missing before. That
test would fail without the previous paragraph.