[WasmFS] Move O_TRUNC handling to avoid crash (#17192)
WasmFS backends like OPFS may reasonably expect that files will only be modified
while open and may be unprepared to modify unopened files. However, when
handling O_TRUNC, we truncated files before opening them in the backend. For the
OPFS backend in particular, this caused a crash because the backend did not yet
have an AccessHandle for the file. Fix the problem by deferring O_TRUNC handling
until after we have asked the backend to open the file.