Remove shared.PYTHON. NFC
This was nothing more than an alias for sys.executable.  While this
is used a fair amount in testing, emscripten itself should not be
forking python subprocesses itself, but instead using the launcher
scripts.

There is one place in emscripten where we do fork a python subprocess
directly (when we run tools/wasm-sourcemap.py) and I'm not sure its
worth adding a wrapper for that minor tool.  For that one case I've
added `-E` to match the changes we are planning to the launcher
scripts: #16736