[WasmFS] Support multiple readers in the OPFS backend
Support multiple readers by using the Blob API instead of AccessHandles to
access files that are opened for reading only. This may come with a slight
performance penalty for accessing read-only files, but the previous behavior can
be recovered by opening file for both reading and writing.

When a file opened only for reading is additionally opened for writing, the
backend gracefully upgrades the access mode to use AccessHandles. Going the
other direction and downgrading the access mode back to Blobs when the last
writable file descriptor is closed is left as a future improvement.

Although it would be possible to support truncation of files opened in the Blob
access mode, we currently disallow it. Truncating (or otherwise modifying) a
file invalidates all Blobs that refer to it, so to support truncation we would
need to have a system for automatically refreshing Blobs for modified files.
That system would ideally also prevent concurrent in-flight operations on the
same file from observing the invalidated Blob.

For the same reason, if one tab writes to an OPFS file via AccessHandles, it
will invalidate the Blobs on all other tabs that have the file open for reading.
This will currently cause all subsequent reads on those other tabs to fail and
the only way to recover is for those tabs to close and reopen the affected
files. In the future it may be worth trying to automatically refresh the Blob
and retry when these errors occur, but in general it is not possible to prevent
the writer tab from starving out the reader tabs.