Create option for easy exception printing
After #17064 and #17157 exception message printing is working, but it
still requires adding a lot of functions to `EXPORTED_FUNCTIONS` and
`DEFAULT_LIBRARY_FUNCS_TO_INLCLUDE`, which is difficult to use and
requires users to know unnecessary library internals:
https://github.com/emscripten-core/emscripten/blob/8c0fe77d8946a7ec2f73dfa74779ac957dd24530/tests/test_core.py#L1639-L1643

This adds `EXCEPTION_PRINTING_SUPPORT` option, which adds necessary
symbols to those settings for easier use. This also exports functions
with which you can get the C++ tag and the thrown value from
`WebAssembly.Exception` object.

Fixes #17114.