Handle memory exceptions errors in MEMFS
When memory allocation exceptions occur in MEMFS from a thread, it causes the thread to fail without recovery (see #16548 and #11496 for details).  Instead we catch the exception at the source and replace it with an ERRNO exception, which is handled appropriately.  This also works better for non-threaded code.  Other exception values to consider: EDQUOT or ENOSPC.