[WasmFS] /dev/null and directory mtime updates
Implement all the functionality needed to pass wasmfs.test_stat except for the
ability to follow symlinks. In particular, implement /dev/null and update
directory mtimes when their contents are changed.

Update WasmFS to unconditionally return 0 as the st_rdev to bring it in line
with the legacy file system on normal files, although the legacy file system
reports other values for special files.