Avoid keeping module file open in tools/webassembly.py API
These was a bug in my imlemenation of `@cache` that I added in #17255
where it was keeping the underlying wasm file open and never actually
closing it.  This is because I was using the incoming arguments as the
hash key instead of their hash value.

The bug only manifested on windows where open files are locked in some
way prevented futher actions from modifying them and resulting is a
strange error:

```llvm-objcopy.exe: error: permission denied```