Fix python metadata extraction with pthreads (#17263)
We have to jump through a lot more hoops to find the EM_JS segment
data when the segments are passive.  In this case the address of each
segment is not in the segment information, but embedded in the
`__wasm_init_memory` code which gets called from `_start`.  We have to
do more work here partially decode this function, but its still less
than binaryen is doing.

Fixes: #16573