Remove special handling of stackSave/stackRestore/stackAlloc
There is not different between how these symbols are handled under the
standard an minimal runtime (unlike the other symbols that are added
separately there).

I'm working changes to remove all the different in the library vs
runtime split that exist between minimal runtime and regular runtime.
For most symbols this means moving them to JS library functions but for
these three they are already handled the same in both runtimes.