Remove redundant alias handling in JS compiler. NFC
This alias handling code was not active.  You can see this from the
fact that the `deps` variable is declared const and then re-assigned
in the case of an alias, which would have been a runtime error if it
ever happended.

Aliased are instead handed in modules.js.  See `apply synonyms` in that
file.