Don't always export writeStackCookie/checkStackCookie. NFC
At some point these were used via the Module object but that is no
longer the case so there is no need to export them.

See #12630 for when the last such usage was removed.

The code size regressions are only in ASSERTIONS builds and they are due
to the assertions being present to check for unexported runtime methods.