Allow DWARF with source map
This allows generation of DWARF info and source map at the same time.
This adds an internal setting, `GENERATE_DWARF`, which is a counterpart
of existing `GENERATE_SORUCE_MAP`, because with only `DEBUG_LEVEL` we
now don't have a way to tell whether DWARF info is explicitly requested
or not, because we set `DEBUG_LEVEL` to 3 when source map is requested.

This adds some simple utility functions to `test_other.py` for DWARF and
source map verification. Also this enables `try_delete` to delete a list
of files.

Fixes #16699.