tests: Avoid duplicating/inlining browser_reporting.js within reftest.js
This file was duplicated/inlined into reftest.js for the case where
the module is running on worker (e.g. proxy-to-worker) and therefore
the copy that gets included via `--pre-js` is not accessible

However, we can simplify this by simply directly including
`browser_reporting.js` when we use it in this way.