Use regular `-I` include flag for SDL paths
When SDL it used on the desktop the SDL include paths are injected
in this way:

```
$ sdl-config  --cflags
-I/usr/include/SDL -D_GNU_SOURCE=1 -D_REENTRANT
$ sdl2-config  --cflags
-I/usr/include/SDL2 -D_REENTRANT
```

Its also the way to add include paths for all the other ports.

As part of this I also refactored the generation of C flags such that
we don't pass `-I` flags when building `.i`/`.ii` files.  These files
are considered to be already pre-processed and clang warns if you pass
`-I` flags when compiling them.