Remove unused `calledMain` global. NFC
This was only used by `proxyWorker.js` so move it there and use the
existing `postRun` mechanism to set it.

The motivation for this change is that it simplifies `callMain` in a way
that should allow us to use `callUserCallback` there rather than the
current/local try/catch.

`calledMain` is not references anywhere else in the codebase or in any
documentation so I think removing it should be safe.