Add SDL_config file for emscripten (#17487)
I noticed that we lacked such a thing while working on #17475.  It turns
out that system header are technically allowed to redefine macros
without generating a warning, but when I made the SDL header tree into a
normal `-I` path I noticed that `M_PI` was being duplicately defined
because we were not defining `HAVE_M_PI`.

This list of defines is taken from `SDL_config_iphoneos.h` and a visual
inspection seems to confirm that we do indeed support all the specified
features.