Rename tests/ to test/. NFC
The name of the directory where tests live is more commonly known as
just `test`.  This matches the convention used by sister projects of
emscripten: llvm, binaryen, and wabt.

This is big change but git is good a tracking renames so anyone with
outstanding changes in this directory should not have any issues.