Return a pair [type, what] from getExceptionMessage (#17386)
As discussed in #17362, this changes `getExceptionMessage` to return a pair, the exception type and the exception `what()` if available. If no exception description is available, the second entry will be undefined. 

This is an API break so we may want to give the new function a different name and keep the old one or something. The behavior of `__get_exception_terminate_message` should be unchanged. This change adds a few extra heap allocations but this is not supposed to be hot code.