Create utils.delete_file and utils.delete_dir in place of tempfiles.try_delete.  NFC (#17512)
The code in `tempfiles.try_delete` went to great length to never fail
and change the permission bits of the things it was trying to delete.

However, in most cases is undesirable and overkill.  We don't want to
silently fail if the delete fails, and we don't want to delete read only
files in most cases.

The exception is in the test suite were we do sometimes create read only
and we want to be able to clean them up.  So I created a test-only
helper called `force_delete_dir`, but even then we don't want to
silently fail so the try/catch was removed here.