Send unhandledrejection events to the main thread (#17535)
We send uncaught exceptions to the main thread where they can be handled in a
central location, but we did not previously send unhandledrejection events that
are created when a promise is rejected with no handler. Add a handler in
worker.js that intercepts unhandledrejection events and throws them as errors so
they are propagated to the main thread in the usual manner.

The way this works varies slightly between Node and the browser, so add tests
for both cases.