Make test_emsymbolizer robust
Currently the addresses are hardcoded, but we have seen them change
multiple times in the last few weeks now. I think it's worthwhile to
make them computed automatically at this point. It'd been easier to use
`wasm-objdump` from wabt, but it is not a dependency of emscripten, so
this uses `llvm-objdump`.