Improve error messages in file_packager.py.
The terms `not be under` and `is below` here were confusing/wrong I
hink.  This error is generated when files are "not contained within"
the current working directory which in mind is the same as "not below".
However the error message was using the term "below" to been "outside
of" which seems wrong to me.

Fixes: #2656