Enable accessing fields that are classes without copying.
Previously, fields that were classes were copied each time they were
accessed from JS. This led to some unexpected behavior (see #6402 and
#17573).

This is a revival of PR #6402. However, I opted to not add a new method
and instead add a new template specialization for `property` for fields
that are classes.

Fixes #6402, #17573