Update test_implict_func for LLVM 16
In LLVM 16, -Wincompatible-function-pointer types has been upgraded to an error
by default. The functional change in this PR is to update test_implicit_func
to incorporate this.
I also refactored the test and captured more precisely the differences between
the messages in C89 vs other modes. The test also no longer links or runs the
result; it didn't seem useful to test for a behavior that is known to be
incorrect, and it speeds up the test.